requirements.txt
torch==2.6.0
pyngrok==7.2.8
pillow==11.1.0
open_clip_torch==2.32.0
numpy==1.26.4
flask==3.1.1
